tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/012.html

<!doctype html>
<title>WebSockets: setting event handlers to ";"</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var events = ['onclose', 'onopen', 'onerror', 'onmessage'];
for (var i = 0; i < events.length; ++i) {
  test(function(t) {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/empty-message');
    ws[events[i]] = ";";
    assert_equals(ws[events[i]], null);
  }, events[i]);
};
</script>
